Rebase from ledgerhq + fix Algorand synchro issue
* Use staging address for Cosmos explorer

* Revert gas estimation base value

The estimation was inaccurate anyway because Gaia doesn't include the
cost of reading from and writing to the state during the simulation, so
the gas estimate is always going to be strongly underevaluated.

And the change here seems also break ledger-live-common in some way, so
better to revert every possible cause of bugs that doesn't bring value

* [ci] test Debug only build

* [ci]filter coin branch name only debug buil available

* [ci]filter coin branch name only debug buil available

* [ci]filter coin branch name only debug buil available

* [ci]filter coin branch name only debug buil available

* [ci] comment failing test

* Create fake explorer for tests

Implement a logic of real explorer to allow synchronization testing
without real data in blockchain.

* Add new ThreadDispatcher based on libuv

* Fix RBF

* [ci] install awscli for coin impl branch

* [ci] update deployment for coin branches

* [ci] update deployment for coin branches

* [ci] add release support

* [ci] bug on artefact's s3 folder

* [Tx] Check if TxUid already exists before insert new Tx

* [tx] delete unuse include and debug variable

* Fix SQL query to remove mempool operations

* [tx] only one commit per tag

* [pr] fix review

* [ci] only on int- branch

* [ci] only on int- branch

+++ .circleci/setup_debian.sh
@@ -16,14 +16,16 @@ echo "========> Java Installed"
 echo $JAVA_HOME
 ls -la /usr/lib/jvm/java-8-openjdk || echo "!!!! java openjdk not found"
 
-apt-get install -y awscli
-echo "========> Install sbt"
-SBT_DEB="sbt-1.3.10.deb"
-curl -L -o $SBT_DEB https://dl.bintray.com/sbt/debian/$SBT_DEB
-dpkg -i $SBT_DEB
-rm $SBT_DEB
-sbt sbtVersion
-apt-get install -y scala
+if [[ "$BUILD_CONFIG" == "Release" || "$CIRCLE_BRANCH" =~ ^int.* ]]; then
+    apt-get install -y awscli
+    echo "========> Install sbt"
+    SBT_DEB="sbt-1.3.10.deb"
+    curl -L -o $SBT_DEB https://dl.bintray.com/sbt/debian/$SBT_DEB
+    dpkg -i $SBT_DEB
+    rm $SBT_DEB
+    sbt sbtVersion
+    apt-get install -y scala
+fi
 
 echo "========> Install C++ dependencies"
 apt-get install -y g++ make
